refactor: service 구현 내용을 cars로 위임함에 따라 테스트 코드 수정
