+++ src/test/java/model/CarsTest.java
@@ -7,10 +7,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.List;
-import java.util.stream.IntStream;
 import org.junit.jupiter.api.Test;
+import support.MovableNumberGenerator;
+import util.NumberGenerator;
 
 class CarsTest {
+    private final NumberGenerator movableNumberGenerator = new MovableNumberGenerator();
 
     @Test
     void 중복된_자동차가_존재하지_않으면_예외가_발생하지_않는다() {
@@ -39,18 +41,37 @@ class CarsTest {
     }
 
     @Test
-    void 가장_많이_전진한_자동차의_위치를_반환한다() {
+    void 단독_우승자를_찾는다() {
         // given
-        Car car1 = 자동차("pobi");
-        Car car2 = 자동차("crong");
-        Cars racingCars = 자동차들(car1, car2);
-        int moveCount = 5;
+        Car car1 = 자동차("prin");
+        Car car2 = 자동차("mark");
+        Car car3 = 자동차("pobi");
+        Cars cars = 자동차들(car1, car2, car3);
+        car1.moveForward(movableNumberGenerator);
 
         // when
-        IntStream.range(0, moveCount).forEach(i -> car1.moveForward());
-        int maxPosition = racingCars.findMaxPosition();
+        List<String> winners = cars.findWinners();
 
         // then
-        assertThat(maxPosition).isEqualTo(moveCount);
+        assertThat(winners).hasSize(1)
+                .contains(car1.getName());
+    }
+
+    @Test
+    void 공동_우승자를_찾는다() {
+        // given
+        Car car1 = 자동차("prin");
+        Car car2 = 자동차("mark");
+        Car car3 = 자동차("pobi");
+        Cars cars = 자동차들(car1, car2, car3);
+        car2.moveForward(movableNumberGenerator);
+        car3.moveForward(movableNumberGenerator);
+
+        // when
+        List<String> winners = cars.findWinners();
+
+        // then
+        assertThat(winners).hasSize(2)
+                .contains(car2.getName(), car3.getName());
     }
 }
